0.4.0 update
In this release effort has gone into making it easier to evaluate
a working version of this project on MBED.

Effort has also gone into simplifying porting. This was to the detriment
of the Arduino wrapper which depended on various strange things to get
around the severe lack of memory on the 328P. The Arduino wrapper has
been removed from the project to save on maintenance. It can be
resurrected if missed, but the memory requirements have increased
slightly so it may no longer work.

There are many breaking changes in this version. If you are updating
from an earlier version you will likely need to make some changes to
your port.

See history.md for more detail.